finished

+++ boats_api/api/boats.js
@@ -3,6 +3,9 @@ const bodyParser = require('body-parser');
 const router = express.Router();
 const ds = require('../lib/datastore');
 const constants = require('../lib/constants');
+const {OAuth2Client} = require('google-auth-library');
+const credentials = require('../lib/oauth_creds');
+const { Z_ASCII } = require('zlib');
 
 const datastore = ds.datastore;
 
@@ -10,18 +13,20 @@ router.use(bodyParser.json());
 
 const c = constants.constants;
 const m = constants.messages;
+const creds = credentials.creds;
+const client = new OAuth2Client(creds.client_id);
 
 
 /* ------------------------------------------------- BEGIN MODEL FUNCTIONS ---------------------------------------- */
 
 // POST a boat
-async function post_boat(name, type, length, public) {
+async function post_boat(name, type, length, public, sub) {
     let key = datastore.key(c.BOAT);
 
     try {
-        const new_boat = {"name":name, "type":type, "length":length, "public": public};
+        const new_boat = {"name":name, "type":type, "length":length, "public": public, "owner": sub};
         await datastore.save({"key":key, "data":new_boat});
-
+        const [boat] = await datastore.get(key);
         return [c.CREATED, ds.fromDatastore(boat)];
 
     } catch (err) {
@@ -32,11 +37,18 @@ async function post_boat(name, type, length, public) {
 }
 
 // GET all the boats in the database
-async function get_all_boats() {
+async function get_all_boats(sub=null) {
     try {
         let q = datastore.createQuery(c.BOAT);
         const entities = await datastore.runQuery(q)
-        return entities[0];
+        let boats;
+        if (sub) {
+            boats = entities[0].filter( boat => boat.owner === sub);
+        } else {
+            boats = entities[0].filter( boat => boat.public);
+        }
+
+        return boats;
 
     } catch (err) {
         console.log('get_all_boats');
@@ -46,24 +58,58 @@ async function get_all_boats() {
 }
 
 // Delete a boat from database by id
-async function delete_boat(id) {
+async function delete_boat(id, sub) {
     try {
         const key = datastore.key([c.BOAT, parseInt(id, 10)]);
+        
+        const [boat] = await datastore.get(key);
 
-        const results = await datastore.delete(key);
+        if (boat && boat.owner === sub) {
+            const results = await datastore.delete(key);
 
-        if (results[0]. indexUpdates === 0){
-            return c.NOT_FOUND;
+            if (results[0]. indexUpdates === 0){
+                return c.FORBIDDEN;
 
+            } else {
+                return c.NO_CONTENT;
+            }
         } else {
-            return c.NO_CONTENT;
+            return c.FORBIDDEN;
         }
+        
 
     } catch (err) {
         console.log('delete_boat');
         console.log(err);
         return c.ERROR;
     }
+}
+
+// verify JWT and send back sub
+async function verify(token) {
+    if (token) {
+        try {
+            const ticket = await client.verifyIdToken({
+                idToken: getToken(token),
+                audience: creds.client_id
+            });
+            const payload = ticket.getPayload();
+            const user_id = payload['sub'];
+            return user_id;
+        } catch (err) {
+            console.log('verify');
+            console.log(err);
+        }
+    } else {
+        return false;
+    }
+    
+}
+
+// parses Authorization header
+function getToken(auth) {
+    const token = auth.split(' ');
+    return token[1];
 }
 
 // send appropriate response
@@ -103,18 +149,7 @@ function handle_response(res, result, extra=null) {
             break;
 
         case c.OK:
-            // send HTML
-            if (extra[1] === 'text/html'){
-                res.status(c.OK).send(createHTML(extra[0]));
-
-            // send JSON
-            } else if (extra[1] === 'application/json') {
-                res.status(c.OK).json(extra[0]);
-
-            // error
-            } else {
-                res.status(c.ERROR).send(message);
-            }
+            res.status(c.OK).send(extra);
             break;
 
         case c.SEE_OTHER:
@@ -123,6 +158,10 @@ function handle_response(res, result, extra=null) {
 
         case c.UNSUPPORTED:
             res.status(c.UNSUPPORTED).send(m.UNSUPPORTED);
+            break;
+
+        case c.UNAUTHORIZED:
+            res.status(c.UNAUTHORIZED).send(m.UNAUTHORIZED);
             break;
     }
 }
@@ -135,74 +174,75 @@ function handle_response(res, result, extra=null) {
 
 // CREATE boat
 router.post('/', async (req, res) => {
-    const accepts = req.accepts(['application/json']);
+    if (req.headers.authorization) {
+        const sub = await verify(req.headers.authorization);    
 
-    // check request content type is correct
-    if (req.get('content-type') !== 'application/json') {
-        handle_response(res, c.UNSUPPORTED);
+        // validate attributes
+        if (sub) {
+            try {
+                // create boat in database
+                const boat = await post_boat(req.body.name, req.body.type, req.body.length, req.body.public, sub);
 
-    // check client response acceptable
-    } else if (!accepts) {
-        handle_response(res, c.NOT_ACCEPTABLE, 'application/json');
+                // handle response
+                handle_response(res, boat[0], boat[1]);
 
-    // validate attributes
-    } else if (req.body.name && req.body.type && req.body.length) {
-
-        try {
-            // create boat in database
-            const boat = await post_boat(req.body.name, req.body.type, req.body.length, req.body.public);
-
-            // handle response
-            handle_response(res, boat[0], boat[1]);
-
-        } catch (err) {
-            console.log('router.post');
-            console.log(err);
-            handle_response(res, c.ERROR);
+            } catch (err) {
+                console.log('router.post');
+                console.log(err);
+                handle_response(res, c.ERROR);
+            }
         }
 
-    // attributes failed
+    // authorization failed
     } else {
-        handle_response(res, c.BAD_REQUEST, m.BAD_REQUEST_ATTR);
+        handle_response(res, c.UNAUTHORIZED, m.UNAUTHORIZED);
     }
 });
 
 // DELETE boat
 router.delete('/:boat_id', async (req, res) => {
-    try {
-        // remove boat from database
-        const result = await delete_boat(req.params.boat_id)
-        handle_response(res, result);
-
-    } catch (err) {
-        console.log('router.delete');
-        console.log(err);
-        handle_response(res, c.ERROR);
+    let sub = await verify(req.headers.authorization);
+    if (sub){
+        try {
+            // remove boat from database
+            const result = await delete_boat(req.params.boat_id, sub);
+            handle_response(res, result);
+    
+        } catch (err) {
+            console.log('router.delete');
+            console.log(err);
+            handle_response(res, c.ERROR);
+        }
+    } else {
+        handle_response(res, c.UNAUTHORIZED);
     }
 });
 
 // GET all boats
 router.get('/', async (req, res) => {
-    // check accepts
-    const accepts = req.accepts(['application/json']);
-    if (!accepts) {
-        handle_response(res, c.NOT_ACCEPTABLE, 'application/json');
+    let sub;
 
+    if (req.headers.authorization) {
+        sub = await verify(req.headers.authorization);
     } else {
-        try {
-            // get boats from database
-            let boats = await get_all_boats();
-            boats = boats.map(ds.fromDatastore);
-            handle_response(res, c.OK, [{"boats": boats}, accepts]);
-
-        } catch (err) {
-            console.log('router.get("/")');
-            console.log(err);
-            handle_response(res, c.ERROR);
-        }
+        sub = null;
     }
+
+    try {
+        // get boats from database
+        let boats = await get_all_boats(sub);
+        boats = boats.map(ds.fromDatastore);
+        handle_response(res, c.OK, boats);
+
+    } catch (err) {
+        console.log('router.get("/")');
+        console.log(err);
+        handle_response(res, c.ERROR);
+    }
+    
 });
 
 /* --------------------------------------------------- END CONTROLLER FUNCTIONS ----------------------------------------- */
 
 module.exports = router;
+module.exports.get_boats = get_all_boats;
